requirements.txt
pytest-playwright

# tutaj  brakuje wersji stabilnych czyli ==1.5.8.